Deus é Fiel

+++ src/main/java/org/example/interfaces/SensorRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface SensorRepository {
     public List<SensorModel> buscarTodos();
-    public SensorModel buscarPorId(long id);
+    public SensorModel buscarPorId(int id);
     public void adicionar(SensorModel sensorModel);
-    public void excluir(long id);
-    public void atualizar(long id, SensorModel sensorModel);
+    public void excluir(int id);
+    public void atualizar(int id, SensorModel sensorModel);
 }
